+++ src/windows/widgets/mod.rs
@@ -1,3 +1,5 @@
+mod layout_button;
 mod workspace_button;
 
+pub use layout_button::*;
 pub use workspace_button::*;